\name{EBarrays-internal}

\alias{emfit0}
\alias{emfit1}
\alias{emfit2}
\alias{emfit3}
\alias{emfitmv}

\title{Internal EBarrays Functions}
\description{
  Internal EBarrays functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
