\name{gould}
\alias{gould}
\non_function{}
\title{A dataset of class matrix}
\usage{data(gould)}
\format{
  The data are originally from Affymetrix chips, subsequently processed
  by dChip and then exported to R for analysis.
}
\description{
  This dataset is part of a dataset from a study of gene expression
  patterns of mammary epithelial cells in a rat model of breast cancer,
  consisting of 5000 genes in 4 biological conditions; 10 arrays total. 
}
\source{
  Dr. M.N. Gould's laboratory in UW-Madison
}
\examples{
data(gould)
}
\keyword{datasets}
