\name{emfit}
\alias{emfit}
\alias{emfit,exprSet,character,ebarraysPatterns-method}
\alias{emfit,exprSet,ebarraysFamily,ebarraysPatterns-method}
\alias{emfit,ExpressionSet,character,ebarraysPatterns-method}
\alias{emfit,ExpressionSet,ebarraysFamily,ebarraysPatterns-method}
\alias{emfit,matrix,character,ebarraysPatterns-method}
\alias{emfit,matrix,ebarraysFamily,ebarraysPatterns-method}
\alias{show,ebarraysEmFit-method}
\alias{ebarraysEmFit-class}
\title{ Implements EM algorithm for gene expression mixture model }
\description{
  Implements the EM algorithm for gene expression mixture model
}
\synopsis{
emfit(data,
      family,
      hypotheses,
      ...)
}
\usage{
emfit(data,
      family,
      hypotheses,
      theta.init = NULL, p.init = NULL,
      num.iter = 20,
      verbose = getOption("verbose"),
      trace = TRUE,
      optim.control = list())
}
\arguments{
  \item{data}{ a matrix or an an object of class \code{"exprSet"} or \code{"ExpressionSet"}}
  \item{family}{
    an object of class ``ebarraysFamily'' or a characterstring which can
    be coerced to one. Currently, only the characters "GG" and "LNN" are
    valid. Other families can be supplied by constructing them
    explicitly.
  }
  \item{hypotheses}{
    an object of class ``ebarraysPatterns'' representing the hypotheses
    of interest. Such patterns can be generated by the function
    \code{\link{ebPatterns}}
  }
  \item{theta.init}{
    initial values of theta. Should be appropriate for the corresponding
    \code{family}. If omitted, defaults are calculated based on the
    \code{family}.
  }
  \item{p.init}{
    initial values of mixture probabilities. If omitted, all hypotheses
    are initially taken as equally likely.
  }
  \item{num.iter}{ number of EM iterations}
  \item{verbose}{ logical or numeric (0,1,2) indicating desired level of
    information printed for the user}
  \item{trace}{
    whether the result should have slots \code{thetaTrace} and
    \code{probTrace}, matrices giving values for all iterations
  }
  \item{optim.control}{
    list passed unchanged to \code{\link{optim}} for finer control
  }
}

\note{
  \code{emfit} is a generic function dispatched on the arguments
  \code{data}, \code{family} and \code{hypotheses}.
}

\value{
  an object of formal class ``ebarraysEmFit'' with the following slots

  \item{\code{family}}{
    the \code{family} passed to \code{emfit}
  }
  \item{\code{hypotheses}}{
    the \code{hypotheses} passed to \code{emfit}
  }
  \item{\code{thetaEst}}{
    numeric vector holding the parameter estimates
  }
  \item{\code{probEst}}{
    numeric vector holding estimates of the mixture probabilities
  }
  \item{\code{thetaTrace}}{
    optional matrix holding intermediate values of \code{thetaEst}
  }
  \item{\code{probTrace}}{
    optional matrix holding intermediate values of \code{probEst}
  }

  The \code{show()} method can be used to summarize the object.  It can
  also be used to generate posterior probabilities using
  \code{\link{postprob}}
}

\references{

  UW Madison Biostatistics and Medical Informatics Technical Report
  166, ``On parametric empirical Bayes methods for comparing multiple
  groups using replicated gene expression profiles,'' by
  C.M. Kendziorski, M.A. Newton, H. Lan, and M.N. Gould.  See
  \url{http://www.biostat.wisc.edu/~kendzior/}
}
\author{ Christina Kendziorski, Michael Newton and Deepayan Sarkar }
\seealso{
  \code{\link{ebPatterns}}, \code{\link{ebarraysFamily-class}}
}
\examples{
data(sample.exprSet.1) ## from Biobase
patterns <- ebPatterns(c("1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1",
                         "1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2"))
gg.fit <- emfit(data = sample.exprSet.1, family = "GG", hypotheses = patterns,
                verbose = TRUE)
##                theta.init = c(10, 1, 1), p.init = c(.99, .01))
show(gg.fit)
}
\keyword{models}

